requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aioice==0.9.0
aiortc==1.10.1
aiosignal==1.3.2
altair==4.2.2
annotated-types==0.7.0
anyio==4.8.0
async-timeout==4.0.3
attrs==25.1.0
audio-recorder-streamlit==0.0.10
av==13.1.0
blinker==1.9.0
bokeh==3.6.3
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.1
cycler==0.12.1
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.7.0
elevenlabs==1.51.0
entrypoints==0.4
exceptiongroup==1.2.2
fonttools==4.56.0
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.44
google-crc32c==1.6.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
ifaddr==0.2.0
Jinja2==3.1.5
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyboard==0.13.5
kiwisolver==1.4.8
langchain==0.3.18
langchain-community==0.3.17
langchain-core==0.3.35
langchain-openai==0.3.5
langchain-text-splitters==0.3.6
langchainhub==0.1.21
langgraph==0.2.72
langgraph-checkpoint==2.0.15
langgraph-sdk==0.1.51
langsmith==0.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.0
mdurl==0.1.2
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.26.0
numpy==1.26.4
openai==1.63.0
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pillow==11.1.0
propcache==0.2.1
protobuf==5.29.3
psycopg2==2.9.10
pyarrow==19.0.0
PyAudio==0.2.14
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydeck==0.9.1
pydub==0.25.1
pyee==12.1.1
Pygments==2.19.1
pylibsrtp==0.11.0
PyMySQL==1.1.1
pyOpenSSL==25.0.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.13.1
SQLAlchemy==2.0.38
streamlit==1.42.0
streamlit-audiorec==0.1.3
streamlit-audiorecorder==0.0.6
streamlit-bokeh-events==0.1.2
streamlit-webrtc==0.47.9
tenacity==9.0.0
tiktoken==0.9.0
toml==0.10.2
toolz==1.0.0
tornado==6.4.2
tqdm==4.67.1
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
watchdog==6.0.0
webrtcvad==2.0.10
websockets==14.2
xyzservices==2025.1.0
yarl==1.18.3
zstandard==0.23.0
